Update bfloat16.jl
"The tests fail due to an unrelated constructor not inferring the return type properly prior to 1.9."

+++ src/bfloat16.jl
@@ -196,7 +196,7 @@ if llvm_arithmetic
         end
     end
 else
-    BFloat16(x::Integer) = convert(BFloat16, convert(Float32, x))
+    BFloat16(x::Integer) = convert(BFloat16, convert(Float32, x)::Float32)
 end
 # TODO: optimize
 BFloat16(x::UInt128) = convert(BFloat16, Float64(x))
